service/requirements.txt
numpy
pandas
scipy
hmmlearn
joblib
bleak
websockets
pytest-asyncio>=0.22.0